use renames fixture file
